Bug 553548: Update to LSP4J 0.9.0 which has DAP 1.36 and 1.37 changes
The 1.36 changes are API breaking changes and that is what the changes
here include.

+++ org.eclipse.lsp4e.debug/src/org/eclipse/lsp4e/debug/debugmodel/DSPBreakpointManager.java
@@ -173,7 +173,7 @@ public class DSPBreakpointManager implements IBreakpointManagerListener, IBreakp
 			List<SourceBreakpoint> sourceBreakpoints = targetBreakpoints.computeIfAbsent(source,
 					s -> new ArrayList<>());
 			SourceBreakpoint sourceBreakpoint = new SourceBreakpoint();
-			sourceBreakpoint.setLine((long) lineNumber);
+			sourceBreakpoint.setLine(lineNumber);
 			sourceBreakpoints.add(sourceBreakpoint);
 		}
 	}
@@ -186,7 +186,7 @@ public class DSPBreakpointManager implements IBreakpointManagerListener, IBreakp
 			IPath location = resource.getLocation();
 			String path = location.toOSString();
 			String name = location.lastSegment();
-			long lineNumber;
+			int lineNumber;
 			try {
 				lineNumber = lineBreakpoint.getLineNumber();
 			} catch (CoreException e) {
@@ -215,7 +215,7 @@ public class DSPBreakpointManager implements IBreakpointManagerListener, IBreakp
 
 			Source source = entry.getKey();
 			List<SourceBreakpoint> bps = entry.getValue();
-			Long[] lines = bps.stream().map(SourceBreakpoint::getLine).toArray(Long[]::new);
+			Integer[] lines = bps.stream().map(SourceBreakpoint::getLine).toArray(Integer[]::new);
 			SourceBreakpoint[] sourceBps = bps.toArray(new SourceBreakpoint[bps.size()]);
 
 			SetBreakpointsArguments arguments = new SetBreakpointsArguments();
